+++ src/main/java/org/folio/persist/InstanceRepository.java
@@ -40,6 +40,24 @@ public class InstanceRepository extends AbstractRepository<Instance> {
     super(postgresClient(context, okapiHeaders), INSTANCE_TABLE, Instance.class);
   }
 
+  public Future<Boolean> doesInstanceExistBySubjectSourceId(String sourceId) {
+    var sql = new StringBuilder("SELECT exists ( ");
+    sql.append("SELECT 1 FROM ");
+    sql.append(INSTANCE_TABLE);
+    sql.append(String.format(" WHERE jsonb @> '{\"subjects\": [{\"sourceId\": \"%s\"}] }' ", sourceId));
+    sql.append(") as record_exists");
+
+    return postgresClient.execute(sql.toString())
+      .map(rowSet -> {
+        if (rowSet != null && rowSet.iterator().hasNext()) {
+          Row row = rowSet.iterator().next();
+          return row.getBoolean("record_exists"); // Extracts the boolean from the result
+        }
+        return false; // Default to false if no results
+      });
+  }
+
+
   public Future<RowStream<Row>> getAllIds(SQLConnection connection) {
     return postgresClientFuturized.selectStream(connection,
       "SELECT id FROM " + postgresClientFuturized.getFullTableName(INSTANCE_TABLE));