+++ src/main/java/org/folio/rest/support/SubjectUtil.java
@@ -4,6 +4,7 @@ import static io.vertx.core.Future.succeededFuture;
 import static org.folio.rest.jaxrs.resource.SubjectSources.PostSubjectSourcesResponse.respond422WithApplicationJson;
 import static org.folio.rest.support.ResponseUtil.SOURCE_CANNOT_BE_FOLIO;
 import static org.folio.rest.support.ResponseUtil.SOURCE_CANNOT_BE_UPDATED_AT_NON_ECS;
+import static org.folio.rest.support.ResponseUtil.SOURCE_CANNOT_BE_DELETED_USED_BY_INSTANCE;
 import static org.folio.rest.support.ResponseUtil.SOURCE_CONSORTIUM_CANNOT_BE_APPLIED;
 import static org.folio.rest.support.ResponseUtil.SOURCE_FOLIO_CANNOT_BE_UPDATED;
 import static org.folio.rest.tools.utils.ValidationHelper.createValidationErrorMessage;
@@ -12,6 +13,7 @@ import io.vertx.core.Future;
 import java.util.Map;
 import java.util.Optional;
 import javax.ws.rs.core.Response;
+import org.folio.persist.InstanceRepository;
 import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.SubjectSource;
 import org.folio.services.consortium.ConsortiumService;
@@ -61,6 +63,17 @@ public final class SubjectUtil {
         });
     }
     return Future.succeededFuture(Optional.empty());
+  }
+
+  public static Future<Optional<Errors>> validateSubjectSourceDelete(String sourceId,
+                                                                     InstanceRepository instanceRepository) {
+    return instanceRepository.doesInstanceExistBySubjectSourceId(sourceId)
+      .compose(exists -> {
+        if (exists) {
+           return Future.succeededFuture(getValidationErrorMessage(sourceId, SOURCE_CANNOT_BE_DELETED_USED_BY_INSTANCE));
+        }
+        return Future.succeededFuture(Optional.empty());
+      });
   }
 
   public static Future<Response> sourceValidationError(Errors errors) {